add registration of entry points used by .C calls
Missing registration is given as a reason for archiving the
package on 2018-01-27 (version 0.6). This commit adds file
init.c in the package src/ (which needs git add -f to
override .gitignore) and modifies src/Makevars and NAMESPACE
to trigger registration.

The change was built upon the BLPackage following INSTALL
instructions for making an R package.